[WIP] Integrate InputMethodService (T9Vietnamese)
Make engine work in different thread
