+++ app/src/main/java/com/vutrankien/t9vietnamese/android/MainActivity.kt
@@ -67,7 +67,13 @@ class MainActivity : Activity(), MVPView {
         setContentView(R.layout.main)
         val kbView = findViewById<KeyboardView>(R.id.dialpad)
         kbView.keyboard = T9Keyboard(this, R.xml.t9)
-        kbView.setOnKeyboardActionListener(KeyboardActionListener(logFactory,scope,eventSource))
+        kbView.setOnKeyboardActionListener(
+            KeyboardActionListener(
+                logFactory,
+                scope,
+                eventSource
+            )
+        )
         val recyclerView = findViewById<RecyclerView>(R.id.recycler_view)
         recyclerView.layoutManager = LinearLayoutManager(this)
         recyclerView.adapter = wordListAdapter
@@ -115,49 +121,6 @@ class MainActivity : Activity(), MVPView {
         log.d("onCandidateClick()")
         //engine.flush()
         (view as TextView).text = ""
-    }
-
-    private class KeyboardActionListener(
-        logFactory: LogFactory,
-        val scope: CoroutineScope,
-        val eventSource: Channel<EventWithData<Event, Key>>
-    ) : KeyboardView.OnKeyboardActionListener {
-        private val log: LogFactory.Log = logFactory.newLog("KeyboardActionListener")
-        override fun swipeRight() {
-            log.d("swipeRight")
-        }
-
-        override fun onPress(primaryCode: Int) {
-            log.d("onPress")
-        }
-
-        override fun onRelease(primaryCode: Int) {
-            log.d("onRelease")
-        }
-
-        override fun swipeLeft() {
-            log.d("swipeLeft")
-        }
-
-        override fun swipeUp() {
-            log.d("swipeUp")
-        }
-
-        override fun swipeDown() {
-            log.d("swipeDown")
-        }
-
-        override fun onKey(primaryCode: Int, keyCodes: IntArray?) {
-            log.d("onKey:$primaryCode,${keyCodes?.joinToString()}")
-        }
-
-        override fun onText(text: CharSequence?) {
-            log.d("onText:$text")
-            if(text == null) return
-            scope.launch {
-                eventSource.send(Event.KEY_PRESS.withData(Key.fromNum(text[0])))
-            }
-        }
     }
 
     fun onBtnClick(view: View) {
