+++ lib/src/main/java/com/vutrankien/t9vietnamese/lib/Presenter.kt
@@ -1,23 +1,19 @@
 package com.vutrankien.t9vietnamese.lib
 
 import com.vutrankien.t9vietnamese.engine.T9Engine
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import kotlin.system.measureTimeMillis
 
 class Presenter(
     private val engineSeed: Lazy<Sequence<String>>,
     private val engine: T9Engine,
     private val env: Env,
-    private val lg: LogFactory
+    lg: LogFactory
 ) {
     private val log = lg.newLog("Presenter")
     private lateinit var view: View
-    internal var typingState: TypingState =
-        TypingState.Init(this)
-        set(value) {
-            log.i("Presenter:TypingState:change: $field -> $value")
-            field = value
-        }
 
     fun attachView(view: View) {
         view.scope.launch {
@@ -47,8 +43,10 @@ class Presenter(
                     view.showKeyboard()
                 }
                 Event.KEY_PRESS -> {
-                    engine.push(eventWithData.data ?:
-                    throw IllegalStateException("UI KEY_PRESS event with null data!"))
+                    withContext(Dispatchers.Default) {
+                        engine.push(eventWithData.data ?:
+                        throw IllegalStateException("UI KEY_PRESS event with null data!"))
+                    }
                 }
             }
         }
@@ -64,41 +62,4 @@ class Presenter(
         }
     }
 
-    sealed class TypingState(lg: LogFactory) {
-        protected val log = lg.newLog("TypingState")
-        open suspend fun keyPress(engine: T9Engine, key: Key) {
-            throw IllegalStateException("${javaClass.name}.keyPress($key)")
-        }
-
-        override fun toString(): String {
-            return javaClass.simpleName
-        }
-
-        class Init(private val presenter: Presenter) : TypingState(presenter.lg) {
-            override suspend fun keyPress(engine: T9Engine, key: Key) {
-                presenter.typingState =
-                    Typing(
-                        presenter,
-                        engine
-                    )
-            }
-        }
-
-        class Typing(private val presenter: Presenter, engine: T9Engine) : TypingState(presenter.lg) {
-            override suspend fun keyPress(engine: T9Engine, key: Key) {
-                log.d("keyPress:$key")
-                engine.push(key)
-            }
-        }
-
-        class Confirmed(presenter: Presenter, result: Set<String>) : TypingState(presenter.lg) {
-            init {
-                presenter.onTypingConfirmed(result)
-            }
-        }
-    }
-
-    private fun onTypingConfirmed(result: Set<String>) {
-        view.showCandidates(result)
-    }
 }
